Some cleanup and tweaks to bringup launch files.
