+++ ateam_bringup/ateam_bringup/__init__.py
@@ -1,0 +1,5 @@
+from . import substitutions
+
+__all__ = [
+    'substitutions'
+]
