--- ateam_bringup/launch/bringup_competition.launch.py
@@ -1,94 +1,0 @@
-# Copyright 2021 A Team
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-import os
-import launch
-from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from launch.launch_description_sources import FrontendLaunchDescriptionSource
-from launch.actions import DeclareLaunchArgument
-from launch_ros.actions import Node
-from ament_index_python.packages import get_package_share_directory
-
-
-def remap_indexed_topics(pattern_from, pattern_to):
-    return list(
-        zip(
-            [pattern_from + str(i) for i in range(16)],
-            [pattern_to + str(i) for i in range(16)],
-        )
-    )
-
-
-def generate_launch_description():
-    ssl_vision_port_value = LaunchConfiguration("ssl_vision_port")
-    ssl_vision_port_arg = DeclareLaunchArgument(
-        "ssl_vision_port", default_value="10006"
-    )
-
-    ateam_bringup_path = os.path.join(
-        get_package_share_directory("ateam_bringup"), "launch"
-    )
-    game_controller_bridge_launch = launch.actions.IncludeLaunchDescription(
-        FrontendLaunchDescriptionSource(
-            [ateam_bringup_path, "/game_controller_nodes.launch.xml"]
-        )
-    )
-    autonomy_launch = launch.actions.IncludeLaunchDescription(
-        FrontendLaunchDescriptionSource([ateam_bringup_path, "/autonomy.launch.xml"]),
-        launch_arguments={"ssl_vision_port": ssl_vision_port_value}.items(),
-    )
-
-    radio_bridge_name = "radio_bridge"
-    radio_bridge = Node(
-        package="ateam_radio_bridge",
-        executable="radio_bridge_node",
-        name=radio_bridge_name,
-        parameters=[{"net_interface_address": "172.16.1.10"}],
-        remappings=remap_indexed_topics(
-            "/" + radio_bridge_name + "/robot_motion_commands/robot",
-            "/robot_motion_commands/robot",
-        )
-        + remap_indexed_topics(
-            "/" + radio_bridge_name + "/robot_feedback/robot", "/robot_feedback/robot"
-        ),
-    )
-
-    ui_launch = launch.actions.IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            PathJoinSubstitution(
-                [
-                    get_package_share_directory("ateam_ui"),
-                    "launch",
-                    "ateam_ui_launch.py",
-                ]
-            )
-        )
-    )
-
-    return launch.LaunchDescription(
-        [
-            ssl_vision_port_arg,
-            autonomy_launch,
-            game_controller_bridge_launch,
-            radio_bridge,
-            ui_launch,
-        ]
-    )
