+++ ateam_bringup/launch/ssl_grsim.launch.xml
@@ -1,14 +1,6 @@
 <launch>
   <arg name="headless" default="true"/>
-  
-  <arg name="ssl_sim_radio_ip" default="127.0.0.1"/>
-  <arg name="ssl_sim_radio_port" default="10301"/>
 
   <executable if="$(var headless)" cmd="docker run --net host --rm robocupssl/grsim" output="screen" respawn="True"/>
   <executable unless="$(var headless)" cmd="docker run --net host -eVNC_PASSWORD=vnc -eVNC_GEOMETRY=1920x1080 --rm robocupssl/grsim vnc" output="screen" respawn="True"/>
-
-  <node name="radio_bridge" pkg="ateam_ssl_simulation_radio_bridge" exec="ssl_simulation_radio_bridge_node" respawn="True">
-    <param name="ssl_sim_radio_ip" value="$(var ssl_sim_radio_ip)"/>
-    <param name="ssl_sim_radio_port" value="$(var ssl_sim_radio_port)"/>
-  </node>
 </launch>
